Restructure MongoDB store to prepare MongoDB source support
